+++ package.nls.json
@@ -20,6 +20,7 @@
 			"{Locked='](command:workbench.action.setLogLevel)'}"
 		]
 	},
+	"githubPullRequests.codingAgent.description": "Enables integration with the asynchronous GitHub Coding Agent",
 	"githubPullRequests.remotes.markdownDescription": "List of remotes, by name, to fetch pull requests from.",
 	"githubPullRequests.queries.markdownDescription": "Specifies what queries should be used in the GitHub Pull Requests tree. All queries are made against **the currently opened repos**. Each query object has a `label` that will be shown in the tree and a search `query` using [GitHub search syntax](https://help.github.com/en/articles/understanding-the-search-syntax). The following variables can be used: \n - `${user}` will resolve to the currently logged in user \n - `${owner}` will resolve to the owner of the current repository, ex. `microsoft` in `microsoft/vscode` \n - `${repository}` will resolve to the repository name, ex. `vscode` in `microsoft/vscode` \n - `${today-Nd}`, where `N` is the number of days ago, will resolve to a date, ex. `2025-01-04`. \n\n By default these queries define the categories \"Waiting For My Review\", \"Assigned To Me\" and \"Created By Me\". If you want to preserve these, make sure they are still in the array when you modify the setting.",
 	"githubPullRequests.queries.label.description": "The label to display for the query in the Pull Requests tree",
